requirements.txt
langchain
langchain_openai
langchain_core
langchain_community
python-dotenv
ipykernel
streamlit